+++ src/lowlevel/allocator/MemoryBlock.h
@@ -45,7 +45,7 @@ struct SMemoryTracking
 
 struct SHeaderStack
 {
-	unsigned int size;
+	size_t size;
 	void* frontOffset;
 	void* backOffset;
 };
@@ -70,7 +70,7 @@ class HAL_API MemoryBlock
 		SHeaderStack* GetStackHeader(void* memoryBlock);
 		SHeaderPool* GetPoolHeader(void* memoryBlock);
 		void SetPoolHeader(void* memoryBlock, size_t next);
-		uint32 CalculateSize(void* position, size_t size, UseHeader header = INVISION_USE_NO_HEADER,
+		size_t CalculateSize(void* position, size_t size, UseHeader header = INVISION_USE_NO_HEADER,
 			MemoryTracking memTracking = INVISION_DEFAULT_MEMORY_TRACKING,
 			BoundsChecking boundsChecking = INVISION_NO_BOUNDS_CHECKING);
